requirements.txt
asn1crypto==1.2.0
beautifulsoup4==4.8.1
bootstrap4==0.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.7
defusedxml==0.6.0
Django==2.2.1
django-crispy-forms==1.9.1
django-easy-pdf==0.1.1
django-extensions==2.2.9
et-xmlfile==1.0.1
html5lib==1.0.1
idna==2.8
jdcal==1.4.1
oauthlib==3.1.0
openpyxl==3.0.3
Pillow==6.2.0
pycparser==2.19
PyMySQL==0.9.3
PyPDF2==1.26.0
python3-openid==3.1.0
pytz==2019.1
reportlab==3.5.42
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
virtualenv==16.7.5
webencodings==0.5.1
wincertstore==0.2
xhtml2pdf==0.2.4
